pyproject.toml
[project]
name = "aiueoserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "cryptography>=46.0.2",
    "django==4.2",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "mysqlclient>=2.2.7",
    "pillow>=11.3.0",
    "pyjwt>=2.10.1",
    "redis>=6.4.0",
    "uvicorn[standard]>=0.37.0",
]

[[tool.uv.index]]
url="https://pypi.tuna.tsinghua.edu.cn/simple"
default=true
